# Eliminate a redundant data copy in H3 framer