@bryancall 2023-01-23
This is odd, I am seeing worse performance when benchmarking this patch:

Without the patch:
```
finished in 7.23s, 69188.73 req/s, 78.12MB/s
requests: 500000 total, 500000 started, 500000 done, 500000 succeeded, 0 failed, 0 errored, 0 timeout
status codes: 500000 2xx, 0 3xx, 0 4xx, 0 5xx
traffic: 564.58MB (592000000) total, 73.43MB (77000000) headers (space savings 35.56%), 488.28MB (512000000) data
UDP datagram: 695977 sent, 1500234 received
                     min         max         mean         sd        +/- sd
time for request:      100us     30.42ms      1.41ms       536us    72.82%
time for connect:    10.60ms     38.26ms     28.97ms      6.05ms    62.00%
time to 1st byte:    39.49ms     43.89ms     41.77ms       971us    60.00%
req/s           :     692.24      752.59      704.10       16.78    84.00%
```

With the patch:
```
finished in 10.40s, 48079.78 req/s, 54.32MB/s
requests: 500000 total, 500000 started, 500000 done, 500000 succeeded, 0 failed, 0 errored, 0 timeout
status codes: 500000 2xx, 0 3xx, 0 4xx, 0 5xx
traffic: 564.91MB (592355606) total, 73.77MB (77355606) headers (space savings 35.27%), 488.28MB (512000000) data
UDP datagram: 1026377 sent, 2500606 received
                     min         max         mean         sd        +/- sd
time for request:      152us     24.44ms      2.04ms       620us    72.30%
time for connect:     7.15ms     34.88ms     23.30ms      6.57ms    59.00%
time to 1st byte:    24.14ms     39.22ms     34.20ms      3.25ms    53.00%
req/s           :     480.99      536.85      490.17       16.35    84.00%
```