@maskit 2023-01-24
Interesting. It might depend on test condition. I'll benchmark this again.