@bryancall 2023-02-07
Flame graph before this patch:
![flamegraph](https://user-images.githubusercontent.com/216749/217361715-a3e35492-5afe-4147-afb3-f675972a6d8d.svg)